+++ tests/test_workplan.py
@@ -37,19 +37,19 @@ def test_workplan():
     assert workplan.remaining == 0
     assert workplan.duration == 10
     assert workplan.end_date == date(2021, 1, 6)
-    assert workplan.resources_usage == 85.0
+    assert workplan.resources_use == 85.0
     assert workplan.project.resources[0].assigned == ['task3'] * 10
     assert workplan.project.resources[0].assigned_tasks == ['task3']
     assert workplan.project.resources[0].duration == 10
     assert workplan.project.resources[0].end_date == date(2021, 1, 6)
-    assert workplan.project.resources[0].usage == 100
+    assert workplan.project.resources[0].use == 100
     assert workplan.project.resources[1].assigned == (
         (['task2'] * 5) + (['task1'] * 2)
     )
     assert workplan.project.resources[1].assigned_tasks == ['task2', 'task1']
     assert workplan.project.resources[1].duration == 7
     assert workplan.project.resources[1].end_date == date(2020, 12, 31)
-    assert workplan.project.resources[1].usage == 70
+    assert workplan.project.resources[1].use == 70
     assert workplan.project.tasks[0].remaining == 0
     assert workplan.project.tasks[1].remaining == 0
     assert workplan.project.tasks[2].remaining == 0
@@ -107,7 +107,7 @@ def test_build_workplan():
     assert workplan.remaining == 0
     assert workplan.duration == 9
     assert workplan.end_date == date(2021, 1, 5)
-    assert workplan.resources_usage == 94.44444444444444
+    assert workplan.resources_use == 94.44444444444444
     assert workplan.project.resources[0].assigned == (
         (['task3'] * 5) + (['task2'] * 3)
     )
@@ -116,7 +116,7 @@ def test_build_workplan():
     ]
     assert workplan.project.resources[0].duration == 8
     assert workplan.project.resources[0].end_date == date(2021, 1, 4)
-    assert workplan.project.resources[0].usage == 88.88888888888889
+    assert workplan.project.resources[0].use == 88.88888888888889
     assert workplan.project.resources[1].assigned == (
         (['task3'] * 5) + (['task1'] * 2) + (['task2'] * 2)
     )
@@ -125,7 +125,7 @@ def test_build_workplan():
     ]
     assert workplan.project.resources[1].duration == 9
     assert workplan.project.resources[1].end_date == date(2021, 1, 5)
-    assert workplan.project.resources[1].usage == 100.0
+    assert workplan.project.resources[1].use == 100.0
     assert workplan.project.tasks[0].remaining == 0
     assert workplan.project.tasks[1].remaining == 0
     assert workplan.project.tasks[2].remaining == 0
@@ -141,12 +141,12 @@ def test_build_workplan_max_resources():
     assert workplan.remaining == 0
     assert workplan.duration == 10
     assert workplan.end_date == date(2021, 1, 6)
-    assert workplan.resources_usage == 85.0
+    assert workplan.resources_use == 85.0
     assert workplan.project.resources[0].assigned == ['task3'] * 10
     assert workplan.project.resources[0].assigned_tasks == ['task3']
     assert workplan.project.resources[0].duration == 10
     assert workplan.project.resources[0].end_date == date(2021, 1, 6)
-    assert workplan.project.resources[0].usage == 100.0
+    assert workplan.project.resources[0].use == 100.0
     assert workplan.project.resources[1].assigned == (
         (['task2'] * 5) + (['task1'] * 2)
     )
@@ -155,7 +155,7 @@ def test_build_workplan_max_resources():
     ]
     assert workplan.project.resources[1].duration == 7
     assert workplan.project.resources[1].end_date == date(2020, 12, 31)
-    assert workplan.project.resources[1].usage == 70.0
+    assert workplan.project.resources[1].use == 70.0
     assert workplan.project.tasks[0].remaining == 0
     assert workplan.project.tasks[1].remaining == 0
     assert workplan.project.tasks[2].remaining == 0